Refactored each class into separate files.
